requirements.txt
altgraph==0.17.2
ao3-api==2.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
configparser==5.2.0
dearpygui==1.1.3
idna==3.3
lxml==4.7.1
pathlib==1.0.1
pyinstaller-hooks-contrib==2021.4
python-slugify==5.0.2
requests==2.26.0
shiboken6==6.2.2.1
soupsieve==2.3.1
text-unidecode==1.3
urllib3==1.26.7
